Move serde conversions of `OracleConfig` to file
